Simplify block-preview contexts
